furnace/context.py: bind mount '/etc/resolv.conf' from the host
With this solution the '/etc/resolv.conf' of the host can be used inside
the container dynamically, because it is bind mounted before every run.
The bind mount is read-only to avoid the modification of the
'/etc/resolv.conf' of the host machine. This commit prepares the place
for another bind mounts too.